Using SupraSeal C2 (#308)
* feat: add SupraSeal

SupraSeal is an optimized version, which is especially targetd and Filecoin's
SHA256 circuit, but is fully compatible with other proofs. Most parts of that
code are written in C++. Therefore things like GPU selection and some environment
may not apply when the `cuda-supraseal` feature is enabled.

Also the Groth16 parameter handling is fully in C++. To make that integration
as simple as possible, some refactoring made sense.

For example is the public API for Groth16 now exclusively through the functions
in `src/groth16/ext.rs`. This way we can have a specially typed version at
`src/groth16/ext_supraseal.rs`, which is only enabled with the `cuda-supraseal`
feature, which makes things more type safe. Else you would only be able to
catch errors with Groth16 parameter handling at run-time. The functions at
`src/groth16/prover/mod.rs` are no longer public.

The main changes for SupraSeal are in the prover. The function signatures are
compatible, hence only the function bodies contain different implementations,
based on the `cuda-supraseal` feature. To make things easier to maintain, is
the `prover` module split into several files. The common functions are in
`prover/mod.rs`, the current implementation of the prover in `prover/native.rs`
and the new SupraSeal implementation at `prover/supraseal.rs`.

The tests in `groth16/tests.rs` wheren't actually run, du to a misconfiguration
with feature flags. As they are integration tests, they were move to the other
ones in `tests`. Now they run as expected on CI.

BREAKING CHANGE: The `create_proof_batch_priority_nonzk` call was removed as
it wasn't used by anyone.